requirements.txt
google-generativeai
python-dotenv
requests
rich
streamlit
pandas